% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{The application User-Interface}
\usage{
app_ui(request)
}
\arguments{
\item{request}{Shiny request object (used to detect query parameters)}
}
\description{
The application User-Interface
}
