% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ram_model.R
\name{create_ram_model}
\alias{create_ram_model}
\title{Create a RAM Model}
\usage{
create_ram_model(resources, activities)
}
\arguments{
\item{resources}{Data frame from \code{define_resources()}.}

\item{activities}{Data frame from \code{define_activities()}.}
}
\value{
List of class 'ram_model'.
}
\description{
Combines resources and activities into a RAM model object.
}
