% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_extdata}
\alias{get_extdata}
\title{Get Path to Extdata File}
\usage{
get_extdata(filename, mustWork = TRUE)
}
\arguments{
\item{filename}{Name of file under inst/extdata}

\item{mustWork}{Logical; if TRUE throws error if not found}
}
\value{
Full path to the installed extdata file
}
\description{
Returns the installed path to a file bundled in \code{inst/extdata}.
}
