\name{location.formatting}
\alias{location.formatting}

\title{Location Formatting}
\description{Some functions in \code{RAM} expect to find a column with 
provincial/territorial data in the metadata. This data should use the standard 
Canadian provincial/territorial abbreviations:
\itemize{
\item{Alberta - AB}
\item{British Columbia - BC}
\item{Manitoba - MB}
\item{New Brunswick - NB}
\item{Newfoundland and Labrador - NL}
\item{Novia Scotia - NS}
\item{Northwest Territories - NT}
\item{Nunavut - NU}
\item{Ontario - ON}
\item{Prince Edward Island - PE}
\item{Quebec - QC}
\item{Saskatchewan - SK}
\item{Yukon - YT}
}

Support for other locations is not available at this time.
}

\author{
Wen Chen and Joshua Simpson.
}
