% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace-overrides.R
\name{namespace-overrides}
\alias{namespace-overrides}
\title{Namespace Overrides for DT - Shiny}
\description{
We want to use DT's versions of the table-output functions,
not shiny's. This tells R to import all of shiny except those.
}
