% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ram.R
\name{plot_ram}
\alias{plot_ram}
\title{Plot RAM Model Solution (plotly version)}
\usage{
plot_ram(solution)
}
\arguments{
\item{solution}{List returned by \code{solve_ram()}.}
}
\description{
Visualizes the optimal allocation (activities) as an interactive barplot.
}
