% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_get_resource_share_invitations}
\alias{ram_get_resource_share_invitations}
\title{Gets the specified invitations for resource sharing}
\usage{
ram_get_resource_share_invitations(resourceShareInvitationArns,
  resourceShareArns, nextToken, maxResults)
}
\arguments{
\item{resourceShareInvitationArns}{The Amazon Resource Names (ARN) of the invitations.}

\item{resourceShareArns}{The Amazon Resource Names (ARN) of the resource shares.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}
}
\description{
Gets the specified invitations for resource sharing.
}
\section{Request syntax}{
\preformatted{ram$get_resource_share_invitations(
  resourceShareInvitationArns = list(
    "string"
  ),
  resourceShareArns = list(
    "string"
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
