% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the RAM app in specified mode ("solve" or "builder")}
\usage{
run_app(mode = c("solve", "builder"), ...)
}
\arguments{
\item{mode}{Character, one of "solve" or "builder" (default: "solve")}

\item{...}{Additional golem options}
}
\description{
Launch the RAM app in specified mode ("solve" or "builder")
}
