% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app_builder}
\alias{run_app_builder}
\title{Launch the RAM app in builder mode}
\usage{
run_app_builder(...)
}
\arguments{
\item{...}{Additional golem options passed to the app}
}
\description{
Launch the RAM app in builder mode
}
