% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app_solve}
\alias{run_app_solve}
\title{Launch the RAM app in solver mode}
\usage{
run_app_solve(...)
}
\arguments{
\item{...}{Additional golem options passed to the app (via \code{get_golem_options()})}
}
\description{
Launch the RAM app in solver mode
}
