% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ram.R
\name{summary_ram}
\alias{summary_ram}
\title{Summary Table for RAM Model Solution (DT version)}
\usage{
summary_ram(solution)
}
\arguments{
\item{solution}{List returned by \code{solve_ram()}.}
}
\value{
DT datatable with optimal activity levels and objective value.
}
\description{
Returns a DT table summarizing the solution of a solved RAM model.
}
